Clean commented-out code, remove dead functions, simplify array conversions
- em/heterogeneity.py: remove ~200 lines of commented-out math derivations,
  dead logger warnings, unused time import, dangling expressions
- reconstruction/ewald.py: remove broken sphere_sign_hard_assignment (uses
  undefined variables) and unused matvec_experiments
- reconstruction/noise.py, relion_functions.py: remove commented-out plots
- heterogeneity/trajectory.py: consolidate 3 duplicate check_in_bound defs
  into np.clip calls, remove dead commented visualization
- heterogeneity/covariance_core.py: remove old commented kernel selection
- heterogeneity/embedding.py: avoid double conversion astype→asarray
- commands/analyze,compute_state,compute_trajectory: np.asarray(dtype=)
- data_io/dataset.py: remove self-assignment, commented tilt code
- em/sampling.py: remove stale comments

+++ recovar/heterogeneity/embedding.py
@@ -151,7 +151,7 @@ def get_coords_in_basis_and_contrast_3(experiment_dataset, mean_estimate, basis,
     n_units = experiment_dataset.n_units if not force_not_shared_label else experiment_dataset.n_images
 
     # Transfer arrays to GPU once before the batch loop
-    basis = jnp.asarray(basis.astype(experiment_dataset.dtype))
+    basis = jnp.asarray(basis, dtype=experiment_dataset.dtype)
     volume_mask = jnp.asarray(volume_mask, dtype=experiment_dataset.dtype_real)
     mean_estimate = jnp.asarray(mean_estimate, dtype=experiment_dataset.dtype)
     eigenvalues = jnp.asarray(eigenvalues, dtype=experiment_dataset.dtype)
